+++ tests/test_json.c
@@ -71,7 +71,7 @@ CASE("mliJson_init, defaults")
 
 CASE("struct mliJson malloc, and free")
 {
-        size_t json_strlen = 10;
+        uint64_t json_strlen = 10;
         struct mliJson json = mliJson_init();
         CHECK(mliJson_malloc(&json, json_strlen));
         mliJson_free(&json);
